--- backend/vercel.json
@@ -1,24 +1,0 @@
-{
-  "version": 2,
-  "rewrites": [
-    {
-      "source": "/api/split-pdf",
-      "destination": "/api/split_pdf/index.go"
-    },
-    {
-      "source": "/api/merge-pdf",
-      "destination": "/api/merge_pdf.go"
-    }
-  ],
-    "headers": [
-      {
-        "source": "/api/(.*)",
-        "headers": [
-          { "key": "Access-Control-Allow-Credentials", "value": "true" },
-          { "key": "Access-Control-Allow-Origin", "value": "*" },
-          { "key": "Access-Control-Allow-Methods", "value": "GET,OPTIONS,PATCH,DELETE,POST,PUT" },
-          { "key": "Access-Control-Allow-Headers", "value": "X-CSRF-Token, X-Requested-With, Accept, Accept-Version, Content-Length, Content-MD5, Content-Type, Date, X-Api-Version" }
-        ]
-      }
-    ]
-}
