+++ frontend/src/api/request.ts
@@ -1,8 +1,8 @@
 import axios from "axios";
 import { AxiosRequestConfig } from "axios";
 
-export const baseUrl = "https://pdf-edition-v2-backend.vercel.app";
-// export const baseUrl = "http://localhost:8080";
+// export const baseUrl = "https://pdf-edition-v2-backend.vercel.app";
+export const baseUrl = "http://localhost:8080";
 
 export const requestGet = async (endpoint: string, params = {}) => {
   const url = `${baseUrl}${endpoint}`;
